+++ lib/dalli/protocol/connection_manager.rb
@@ -100,13 +100,13 @@ module Dalli
 
       def confirm_ready!
         close if request_in_progress?
-        close_on_fork if fork_detected?
+        reconnect_on_fork if fork_detected?
       end
 
       def confirm_in_progress!
         raise '[Dalli] No request in progress. This may be a bug in Dalli.' unless request_in_progress?
 
-        close_on_fork if fork_detected?
+        reconnect_on_fork if fork_detected?
       end
 
       def close
@@ -218,7 +218,17 @@ module Dalli
         end
       end
 
+      def reconnect_on_fork
+        message = 'Fork detected, re-connecting child process...'
+        Dalli.logger.info { message }
+        # Close socket on a fork and reconnect immediately
+        close
+        establish_connection
+      end
+
       def close_on_fork
+        Dalli.logger.warn 'DEPRECATED: close_on_fork is deprecated and will be removed in a future version. ' \
+                          'Use reconnect_on_fork instead.'
         message = 'Fork detected, re-connecting child process...'
         Dalli.logger.info { message }
         # Close socket on a fork, setting us up for reconnect